vh: mm: Skip is_migrate_cma_page() when a TLB flush is already needed
It is observed that the vh_pagevec_drain() hook can burn a decent bit of
CPU according to perf top, likely because the is_migrate_cma_page() check
is somewhat expensive.

Since the vh_pagevec_drain() hook only attempts to set `*ret` to true, skip
is_migrate_cma_page() when a flush is already needed, as an optimization.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>