uwb: Don't bail out when there's no module owner
The module owner field is NULL when uwb is built-in as opposed to being
built as a module. Since mcps802154_region_register() and
mcps802154_scheduler_register() both mandate having a module owner, this
breaks uwb when it's built-in rather than being built as a module.

Remove the module owner check completely to allow uwb to work built-in.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>